+++ newrelic/exporter.go
@@ -1,28 +1,38 @@
 // Copyright 2019 New Relic Corporation. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package newrelic contains an OpenTelemetry tracing exporter for New Relic.
+// Package newrelic provides an OpenTelemetry exporter for New Relic.
 package newrelic
 
 import (
 	"context"
 	"errors"
-	"go.opentelemetry.io/otel/api/metric"
+	"fmt"
+	"os"
+	"strings"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/global"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
+	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
+	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
+	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
+	"go.opentelemetry.io/otel/sdk/resource"
+	"go.opentelemetry.io/otel/semconv"
+	"go.opentelemetry.io/otel/trace"
 
 	"github.com/newrelic/newrelic-telemetry-sdk-go/telemetry"
 	"github.com/newrelic/opentelemetry-exporter-go/newrelic/internal/transform"
-	metricsdk "go.opentelemetry.io/otel/sdk/export/metric"
-	tracesdk "go.opentelemetry.io/otel/sdk/export/trace"
+	exportmetric "go.opentelemetry.io/otel/sdk/export/metric"
+	exporttrace "go.opentelemetry.io/otel/sdk/export/trace"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
 const (
 	version          = "0.1.0"
 	userAgentProduct = "NewRelic-Go-OpenTelemetry"
 )
-
-// Java implementation:
-// https://github.com/newrelic/newrelic-opentelemetry-java-exporters/tree/master/src/main/java/com/newrelic/telemetry/opentelemetry/export
 
 // Exporter exports OpenTelemetry data to New Relic.
 type Exporter struct {
@@ -35,9 +45,9 @@ var (
 	errServiceNameEmpty = errors.New("service name is required")
 )
 
-// NewExporter creates a new Exporter that exports spans to New Relic.
-func NewExporter(serviceName, apiKey string, options ...func(*telemetry.Config)) (*Exporter, error) {
-	if serviceName == "" {
+// NewExporter creates a new Exporter that exports telemetry to New Relic.
+func NewExporter(service, apiKey string, options ...func(*telemetry.Config)) (*Exporter, error) {
+	if service == "" {
 		return nil, errServiceNameEmpty
 	}
 	options = append([]func(*telemetry.Config){
@@ -53,34 +63,132 @@ func NewExporter(serviceName, apiKey string, options ...func(*telemetry.Config))
 	}
 	return &Exporter{
 		harvester:   h,
-		serviceName: serviceName,
+		serviceName: service,
 	}, nil
 }
 
-var (
-	_ tracesdk.SpanSyncer  = (*Exporter)(nil)
-	_ tracesdk.SpanBatcher = (*Exporter)(nil)
-	_ metricsdk.Exporter   = (*Exporter)(nil)
-)
-
-// ExportSpans exports multiple spans to New Relic.
-func (e *Exporter) ExportSpans(ctx context.Context, spans []*tracesdk.SpanData) {
-	for _, s := range spans {
-		e.ExportSpan(ctx, s)
+// NewExportPipeline creates a new OpenTelemetry telemetry pipeline using a
+// New Relic Exporter configured with default setting. It is the callers
+// responsibility to stop the returned OTel Controller. This function uses the
+// following environment variables to configure the exporter installed in the
+// pipeline:
+//
+//    * `NEW_RELIC_API_KEY`: New Relic Event API key.
+//    * `NEW_RELIC_METRIC_URL`: Override URL to New Relic metric endpoint.
+//    * `NEW_RELIC_TRACE_URL`: Override URL to New Relic trace endpoint.
+//
+// More information about the New Relic Event API key can be found
+// here: https://docs.newrelic.com/docs/apis/get-started/intro-apis/types-new-relic-api-keys#event-insert-key.
+//
+// The exporter will send telemetry to the default New Relic metric and trace
+// API endpoints in the United States. These can be overwritten with the above
+// environment variables. These are useful if you wish to send to our EU
+// endpoints:
+//
+//    * EU metric API endpoint: metric-api.eu.newrelic.com/metric/v1
+//    * EU trace API endpoint: trace-api.eu.newrelic.com/trace/v1
+func NewExportPipeline(service string, traceOpt []sdktrace.TracerProviderOption, cOpt []controller.Option) (trace.TracerProvider, *controller.Controller, error) {
+	apiKey, ok := os.LookupEnv("NEW_RELIC_API_KEY")
+	if !ok {
+		return nil, nil, errors.New("missing New Relic API key")
 	}
+
+	var eOpts []func(*telemetry.Config)
+	if u, ok := os.LookupEnv("NEW_RELIC_METRIC_URL"); ok {
+		eOpts = append(eOpts, func(cfg *telemetry.Config) {
+			cfg.MetricsURLOverride = u
+		})
+	}
+	if u, ok := os.LookupEnv("NEW_RELIC_TRACE_URL"); ok {
+		eOpts = append(eOpts, telemetry.ConfigSpansURLOverride(u))
+	}
+
+	exporter, err := NewExporter(service, apiKey, eOpts...)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// Minimally default resource with a service name. This is overwritten if
+	// another is passed in traceOpt or pushOpt.
+	r := resource.NewWithAttributes(semconv.ServiceNameKey.String(service))
+
+	tp := sdktrace.NewTracerProvider(
+		append([]sdktrace.TracerProviderOption{
+			sdktrace.WithSyncer(exporter),
+			sdktrace.WithResource(r),
+		},
+			traceOpt...)...,
+	)
+
+	pusher := controller.New(
+		processor.New(
+			simple.NewWithExactDistribution(),
+			exporter,
+		),
+		append([]controller.Option{controller.WithResource(r)}, cOpt...)...,
+	)
+	pusher.Start(context.TODO())
+
+	return tp, pusher, nil
 }
 
-// ExportSpan exports a span to New Relic.
-func (e *Exporter) ExportSpan(ctx context.Context, span *tracesdk.SpanData) {
-	if nil == e {
-		return
+// InstallNewPipeline installs a New Relic exporter with default settings
+// in the global OpenTelemetry telemetry pipeline. It is the callers
+// responsibility to stop the returned push Controller. 
+// ## Prerequisites
+// For details, check out the "Get Started" section of [New Relic Go OpenTelemetry exporter](https://github.com/newrelic/opentelemetry-exporter-go/blob/master/README.md#get-started).
+// ## Environment variables
+// This function uses the following environment variables to configure 
+// the exporter installed in the pipeline:
+//    * `NEW_RELIC_API_KEY`: New Relic Insights insert key.
+//    * `NEW_RELIC_METRIC_URL`: Override URL to New Relic metric endpoint.
+//    * `NEW_RELIC_TRACE_URL`: Override URL to New Relic trace endpoint.
+// The exporter will send telemetry to the default New Relic metric and trace
+// API endpoints in the United States:
+// * Traces: https://trace-api.newrelic.com/trace/v1
+// * Metrics: https://metric-api.newrelic.com/metric/v1
+// You can overwrite these with the above environment variables
+// to send data to our EU endpoints or to set up Infinite Tracing.
+// For information about changing endpoints, see [OpenTelemetry: Advanced configuration](https://docs.newrelic.com/docs/integrations/open-source-telemetry-integrations/opentelemetry/opentelemetry-advanced-configuration#h2-change-endpoints).
+
+func InstallNewPipeline(service string) (*controller.Controller, error) {
+	tp, controller, err := NewExportPipeline(service, nil, nil)
+	if err != nil {
+		return nil, err
 	}
-	e.harvester.RecordSpan(transform.Span(e.serviceName, span))
+
+	otel.SetTracerProvider(tp)
+	global.SetMeterProvider(controller.MeterProvider())
+	return controller, nil
+}
+
+var (
+	_ exporttrace.SpanExporter = (*Exporter)(nil)
+	_ exportmetric.Exporter    = (*Exporter)(nil)
+)
+
+// ExportSpans exports span data to New Relic.
+func (e *Exporter) ExportSpans(ctx context.Context, spans []*exporttrace.SpanSnapshot) error {
+	if nil == e {
+		return nil
+	}
+
+	var errs []string
+	for _, s := range spans {
+		if err := e.harvester.RecordSpan(transform.Span(e.serviceName, s)); err != nil {
+			errs = append(errs, err.Error())
+		}
+	}
+
+	if len(errs) > 0 {
+		return fmt.Errorf("export span: %s", strings.Join(errs, ", "))
+	}
+	return nil
 }
 
 // Export exports metrics to New Relic.
-func (e *Exporter) Export(_ context.Context, cps metricsdk.CheckpointSet) error {
-	return cps.ForEach(e, func(record metricsdk.Record) error {
+func (e *Exporter) Export(_ context.Context, cps exportmetric.CheckpointSet) error {
+	return cps.ForEach(e, func(record exportmetric.Record) error {
 		m, err := transform.Record(e.serviceName, record)
 		if err != nil {
 			return err
@@ -90,6 +198,11 @@ func (e *Exporter) Export(_ context.Context, cps metricsdk.CheckpointSet) error 
 	})
 }
 
-func (e *Exporter) ExportKindFor(_ *metric.Descriptor, _ aggregation.Kind) metricsdk.ExportKind {
-	return metricsdk.DeltaExporter
+func (e *Exporter) ExportKindFor(_ *metric.Descriptor, _ aggregation.Kind) exportmetric.ExportKind {
+	return exportmetric.DeltaExportKind
+}
+
+func (e *Exporter) Shutdown(ctx context.Context) error {
+	e.harvester.HarvestNow(ctx)
+	return nil
 }