+++ newrelic/exporter_test.go
@@ -17,14 +17,16 @@ import (
 	"time"
 
 	"github.com/newrelic/newrelic-telemetry-sdk-go/telemetry"
-	"go.opentelemetry.io/otel/api/metric"
-	metricapi "go.opentelemetry.io/otel/api/metric"
-	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/number"
 	"go.opentelemetry.io/otel/sdk/export/trace"
-	"go.opentelemetry.io/otel/sdk/metric/controller/push"
-	integrator "go.opentelemetry.io/otel/sdk/metric/processor/basic"
+	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
+	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	selector "go.opentelemetry.io/otel/sdk/metric/selector/simple"
+	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/semconv"
 )
 
 func TestServiceNameMissing(t *testing.T) {
@@ -38,11 +40,10 @@ func TestServiceNameMissing(t *testing.T) {
 }
 
 func TestNilExporter(t *testing.T) {
-	span := &trace.SpanData{}
+	span := &trace.SpanSnapshot{}
 	var e *Exporter
 
-	e.ExportSpan(context.Background(), span)
-	e.ExportSpans(context.Background(), []*trace.SpanData{span})
+	e.ExportSpans(context.Background(), []*trace.SpanSnapshot{span})
 }
 
 // MockTransport caches decompressed request bodies
@@ -153,27 +154,26 @@ func TestEndToEndTracer(t *testing.T) {
 		t.Fatalf("failed to instantiate exporter: %v", err)
 	}
 
-	traceProvider, err := sdktrace.NewProvider(
+	r := resource.NewWithAttributes(semconv.ServiceNameKey.String(serviceName))
+	tracerProvider := sdktrace.NewTracerProvider(
 		sdktrace.WithBatcher(e, sdktrace.WithBatchTimeout(15), sdktrace.WithMaxExportBatchSize(10)),
+		sdktrace.WithResource(r),
 	)
-	if err != nil {
-		t.Fatalf("failed to instantiate trace provider: %v", err)
-	}
 
-	tracer := traceProvider.Tracer("test-tracer")
+	tracer := tracerProvider.Tracer("test-tracer")
 
-	var decend func(context.Context, int)
-	decend = func(ctx context.Context, n int) {
+	var descend func(context.Context, int)
+	descend = func(ctx context.Context, n int) {
 		if n <= 0 {
 			return
 		}
 		depth := numSpans - n
 		ctx, span := tracer.Start(ctx, fmt.Sprintf("Span %d", depth))
-		span.SetAttributes(label.Int("depth", depth))
-		decend(ctx, n-1)
+		span.SetAttributes(attribute.Int("depth", depth))
+		descend(ctx, n-1)
 		span.End()
 	}
-	decend(context.Background(), numSpans)
+	descend(context.Background(), numSpans)
 
 	// Wait >2 cycles.
 	<-time.After(40 * time.Millisecond)
@@ -217,29 +217,29 @@ func TestEndToEndTracer(t *testing.T) {
 func TestEndToEndMeter(t *testing.T) {
 	serviceName := "opentelemetry-service"
 	type data struct {
-		iKind metric.Kind
-		nKind metric.NumberKind
+		iKind metric.InstrumentKind
+		nKind number.Kind
 		val   int64
 	}
 	instruments := map[string]data{
-		"test-int64-counter":                {metric.CounterKind, metric.Int64NumberKind, 1},
-		"test-float64-counter":              {metric.CounterKind, metric.Float64NumberKind, 1},
-		"test-int64-up-down-counter":        {metric.UpDownCounterKind, metric.Int64NumberKind, 1},
-		"test-float64-up-down-counter":      {metric.UpDownCounterKind, metric.Float64NumberKind, 1},
-		"test-int64-measure":                {metric.ValueRecorderKind, metric.Int64NumberKind, 2},
-		"test-float64-measure":              {metric.ValueRecorderKind, metric.Float64NumberKind, 2},
-		"test-int64-observer":               {metric.ValueObserverKind, metric.Int64NumberKind, 3},
-		"test-float64-observer":             {metric.ValueObserverKind, metric.Float64NumberKind, 3},
-		"test-int64-sum-observer":           {metric.SumObserverKind, metric.Int64NumberKind, 3},
-		"test-float64-sum-observer":         {metric.SumObserverKind, metric.Float64NumberKind, 3},
-		"test-int64-up-down-sum-observer":   {metric.UpDownSumObserverKind, metric.Int64NumberKind, 3},
-		"test-float64-up-down-sum-observer": {metric.UpDownSumObserverKind, metric.Float64NumberKind, 3},
+		"test-int64-counter":                {metric.CounterInstrumentKind, number.Int64Kind, 1},
+		"test-float64-counter":              {metric.CounterInstrumentKind, number.Float64Kind, 1},
+		"test-int64-up-down-counter":        {metric.UpDownCounterInstrumentKind, number.Int64Kind, 1},
+		"test-float64-up-down-counter":      {metric.UpDownCounterInstrumentKind, number.Float64Kind, 1},
+		"test-int64-measure":                {metric.ValueRecorderInstrumentKind, number.Int64Kind, 2},
+		"test-float64-measure":              {metric.ValueRecorderInstrumentKind, number.Float64Kind, 2},
+		"test-int64-observer":               {metric.ValueObserverInstrumentKind, number.Int64Kind, 3},
+		"test-float64-observer":             {metric.ValueObserverInstrumentKind, number.Float64Kind, 3},
+		"test-int64-sum-observer":           {metric.SumObserverInstrumentKind, number.Int64Kind, 3},
+		"test-float64-sum-observer":         {metric.SumObserverInstrumentKind, number.Float64Kind, 3},
+		"test-int64-up-down-sum-observer":   {metric.UpDownSumObserverInstrumentKind, number.Int64Kind, 3},
+		"test-float64-up-down-sum-observer": {metric.UpDownSumObserverInstrumentKind, number.Float64Kind, 3},
 	}
 
 	mockt := &MockTransport{
 		Data: make([]Data, 0, len(instruments)),
 	}
-	e, err := NewExporter(
+	exp, err := NewExporter(
 		serviceName,
 		"apiKey",
 		telemetry.ConfigHarvestPeriod(0),
@@ -256,74 +256,84 @@ func TestEndToEndMeter(t *testing.T) {
 		t.Fatalf("failed to instantiate exporter: %v", err)
 	}
 
-	aggSelector := selector.NewWithExactDistribution()
-	batcher := integrator.New(aggSelector, e)
-	pusher := push.New(batcher, e)
-	pusher.Start()
-
 	ctx := context.Background()
-	meter := pusher.Provider().Meter("test-meter")
+	control := controller.New(
+		processor.New(
+			selector.NewWithInexpensiveDistribution(),
+			exp, // passed as an ExportKindSelector.
+		),
+		// Set collection period longer than this test will run for.
+		controller.WithCollectPeriod(10*time.Second),
+		controller.WithPushTimeout(time.Millisecond),
+		controller.WithPusher(exp),
+	)
 
-	newInt64ObserverCallback := func(v int64) metricapi.Int64ObserverFunc {
-		return func(ctx context.Context, result metricapi.Int64ObserverResult) { result.Observe(v) }
+	if err := control.Start(ctx); err != nil {
+		t.Fatalf("starting controller: %v", err)
 	}
-	newFloat64ObserverCallback := func(v float64) metricapi.Float64ObserverFunc {
-		return func(ctx context.Context, result metricapi.Float64ObserverResult) { result.Observe(v) }
+
+	meter := control.MeterProvider().Meter("test-meter")
+
+	newInt64ObserverCallback := func(v int64) metric.Int64ObserverFunc {
+		return func(ctx context.Context, result metric.Int64ObserverResult) { result.Observe(v) }
+	}
+	newFloat64ObserverCallback := func(v float64) metric.Float64ObserverFunc {
+		return func(ctx context.Context, result metric.Float64ObserverResult) { result.Observe(v) }
 	}
 
 	for name, data := range instruments {
 		switch data.iKind {
-		case metric.CounterKind:
+		case metric.CounterInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64Counter(name).Add(ctx, data.val)
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64Counter(name).Add(ctx, float64(data.val))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64Counter(name).Add(ctx, data.val)
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64Counter(name).Add(ctx, float64(data.val))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
-		case metric.UpDownCounterKind:
+		case metric.UpDownCounterInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64UpDownCounter(name).Add(ctx, data.val)
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64UpDownCounter(name).Add(ctx, float64(data.val))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64UpDownCounter(name).Add(ctx, data.val)
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64UpDownCounter(name).Add(ctx, float64(data.val))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
-		case metric.ValueRecorderKind:
+		case metric.ValueRecorderInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64ValueRecorder(name).Record(ctx, data.val)
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64ValueRecorder(name).Record(ctx, float64(data.val))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64ValueRecorder(name).Record(ctx, data.val)
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64ValueRecorder(name).Record(ctx, float64(data.val))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
-		case metric.ValueObserverKind:
+		case metric.ValueObserverInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64ValueObserver(name, newInt64ObserverCallback(data.val))
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64ValueObserver(name, newFloat64ObserverCallback(float64(data.val)))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64ValueObserver(name, newInt64ObserverCallback(data.val))
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64ValueObserver(name, newFloat64ObserverCallback(float64(data.val)))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
-		case metric.SumObserverKind:
+		case metric.SumObserverInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64SumObserver(name, newInt64ObserverCallback(data.val))
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64SumObserver(name, newFloat64ObserverCallback(float64(data.val)))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64SumObserver(name, newInt64ObserverCallback(data.val))
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64SumObserver(name, newFloat64ObserverCallback(float64(data.val)))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
-		case metric.UpDownSumObserverKind:
+		case metric.UpDownSumObserverInstrumentKind:
 			switch data.nKind {
-			case metric.Int64NumberKind:
-				metricapi.Must(meter).NewInt64UpDownSumObserver(name, newInt64ObserverCallback(data.val))
-			case metric.Float64NumberKind:
-				metricapi.Must(meter).NewFloat64UpDownSumObserver(name, newFloat64ObserverCallback(float64(data.val)))
+			case number.Int64Kind:
+				metric.Must(meter).NewInt64UpDownSumObserver(name, newInt64ObserverCallback(data.val))
+			case number.Float64Kind:
+				metric.Must(meter).NewFloat64UpDownSumObserver(name, newFloat64ObserverCallback(float64(data.val)))
 			default:
 				t.Fatal("unsupported number testing kind", data.nKind.String())
 			}
@@ -332,16 +342,19 @@ func TestEndToEndMeter(t *testing.T) {
 		}
 	}
 
-	// Wait >2 cycles.
-	<-time.After(40 * time.Millisecond)
+	// Flush and stop the conroller.
+	if err := control.Stop(ctx); err != nil {
+		t.Fatalf("stopping controller: %v", err)
+	}
 
-	// Flush and close.
-	pusher.Stop()
-	e.harvester.HarvestNow(ctx)
+	// Flush and stop the exporter.
+	if err := exp.Shutdown(ctx); err != nil {
+		t.Fatalf("shutting down exporter: %v", err)
+	}
 
 	gotMetrics := mockt.Metrics()
 	if got, want := len(gotMetrics), len(instruments); got != want {
-		t.Fatalf("expecting %d spans, got %d", want, got)
+		t.Fatalf("expecting %d metrics, got %d", want, got)
 	}
 	seen := make(map[string]struct{}, len(instruments))
 	for _, m := range gotMetrics {
@@ -353,16 +366,26 @@ func TestEndToEndMeter(t *testing.T) {
 		seen[m.Name] = struct{}{}
 
 		switch want.iKind {
-		case metric.CounterKind:
+		case metric.CounterInstrumentKind:
 			if m.Type != "count" {
 				t.Errorf("metric type for %s: got %q, want \"counter\"", m.Name, m.Type)
+				continue
 			}
 			if got := m.Value.(float64); got != float64(want.val) {
 				t.Errorf("metric value for %s: got %g, want %d", m.Name, m.Value, want.val)
 			}
-		case metric.ValueRecorderKind, metric.ValueObserverKind:
+		case metric.ValueObserverInstrumentKind:
+			if m.Type != "gauge" {
+				t.Errorf("metric type for %s: got %q, want \"gauge\"", m.Name, m.Type)
+				continue
+			}
+			if got := m.Value.(float64); got != float64(want.val) {
+				t.Errorf("metric value for %s: got %g, want %d", m.Name, m.Value, want.val)
+			}
+		case metric.ValueRecorderInstrumentKind:
 			if m.Type != "summary" {
 				t.Errorf("metric type for %s: got %q, want \"summary\"", m.Name, m.Type)
+				continue
 			}
 			value := m.Value.(map[string]interface{})
 			if got := value["count"].(float64); got != 1 {